+++ src/Repository/WLT/ContactRepository.php
@@ -18,6 +18,7 @@
 
 namespace App\Repository\WLT;
 
+use App\Entity\Edu\AcademicYear;
 use App\Entity\Edu\ContactMethod;
 use App\Entity\Edu\Teacher;
 use App\Entity\WLT\Contact;
@@ -80,6 +81,32 @@ class ContactRepository extends ServiceEntityRepository
         return $qb;
     }
 
+    public function getWorkcenterStatsByIdAcademicYearAndFilterQueryBuilder($workcenters, $academicYear, $q)
+    {
+        $qb = $this->getEntityManager()->createQueryBuilder()
+            ->from(Workcenter::class, 'w')
+            ->select('w, COUNT(c), COUNT(c.method)')
+            ->leftJoin(Contact::class, 'c', 'WITH', 'c.workcenter = w')
+            ->leftJoin('w.company', 'co')
+            ->leftJoin('c.teacher', 't')
+            ->andWhere('t.academicYear = :academic_year')
+            ->setParameter('academic_year', $academicYear)
+            ->groupBy('w')
+            ->orderBy('co.name')
+            ->addOrderBy('w.name');
+
+        if ($q) {
+            $qb
+                ->where('co.name LIKE :tq OR w.name LIKE :tq')
+                ->setParameter('tq', '%'. $q . '%');
+        }
+        $qb
+            ->andWhere('w IN (:items)')
+            ->setParameter('items', $workcenters);
+
+        return $qb;
+    }
+
     public function findWorkcentersByTeacher(Teacher $teacher)
     {
         return $this->getEntityManager()->createQueryBuilder()
@@ -88,6 +115,21 @@ class ContactRepository extends ServiceEntityRepository
             ->join(Contact::class, 'c', 'WITH', 'c.workcenter = w AND c.teacher = :teacher')
             ->join('w.company', 'co')
             ->setParameter('teacher', $teacher)
+            ->orderBy('co.name')
+            ->addOrderBy('w.name')
+            ->getQuery()
+            ->getResult();
+    }
+
+    public function findWorkcentersByTeachers($teachers)
+    {
+        return $this->getEntityManager()->createQueryBuilder()
+            ->select('w')
+            ->from(Workcenter::class, 'w')
+            ->join(Contact::class, 'c', 'WITH', 'c.workcenter = w AND c.teacher IN (:teachers)')
+            ->join('w.company', 'co')
+            ->join('c.teacher', 't')
+            ->setParameter('teachers', $teachers)
             ->orderBy('co.name')
             ->addOrderBy('w.name')
             ->getQuery()
@@ -145,6 +187,53 @@ class ContactRepository extends ServiceEntityRepository
             ->getResult();
     }
 
+    public function findByAcademicYearWorkcenterProjectsAndMethods(
+        AcademicYear $academicYear,
+        Workcenter $workcenter,
+        $projects = [],
+        $contactMethods = []
+    ) {
+        $qb = $this->createQueryBuilder('c')
+            ->select('c, p, w, t, m, pe')
+            ->join('c.teacher', 't')
+            ->join('t.person', 'pe')
+            ->leftJoin('c.method', 'm')
+            ->leftJoin('c.projects', 'p')
+            ->leftJoin('c.workcenter', 'w')
+            ->where('t.academicYear = :academic_year')
+            ->setParameter('academic_year', $academicYear)
+            ->andWhere('c.workcenter = :workcenter')
+            ->setParameter('workcenter', $workcenter);
+
+        if ($projects !== []) {
+            $noProject = false;
+            if (($key = array_search(null, $projects, true)) !== false) {
+                unset($projects[$key]);
+                $noProject = true;
+            }
+            $qb
+                ->andWhere('p IN (:projects)' . ($noProject ? ' OR p IS NULL' : ''))
+                ->setParameter('projects', $projects);
+        }
+
+        if ($contactMethods !== []) {
+            $onsite = false;
+            if (($key = array_search(null, $contactMethods, true)) !== false) {
+                unset($contactMethods[$key]);
+                $onsite = true;
+            }
+
+            $qb
+                ->andWhere('m IN (:contact_methods)' . ($onsite ? ' OR m IS NULL' : ''))
+                ->setParameter('contact_methods', $contactMethods);
+        }
+
+        return $qb
+            ->orderBy('c.dateTime')
+            ->getQuery()
+            ->getResult();
+    }
+
     public function getContactMethodStatsByTeacherWorkcenterProjectsAndMethods(
         Teacher $teacher,
         Workcenter $workcenter = null,
@@ -170,6 +259,55 @@ class ContactRepository extends ServiceEntityRepository
                 ->andWhere('c.workcenter = :workcenter')
                 ->setParameter('workcenter', $workcenter);
         }
+
+        if ($projects !== []) {
+            $noProject = false;
+            if (($key = array_search(null, $projects, true)) !== false) {
+                unset($projects[$key]);
+                $noProject = true;
+            }
+            $qb
+                ->andWhere('p IN (:projects)' . ($noProject ? ' OR p IS NULL' : ''))
+                ->setParameter('projects', $projects);
+        }
+
+        if ($contactMethods !== []) {
+            $onsite = false;
+            if (($key = array_search(null, $contactMethods, true)) !== false) {
+                unset($contactMethods[$key]);
+                $onsite = true;
+            }
+
+            $qb
+                ->andWhere('m IN (:contact_methods)' . ($onsite ? ' OR m IS NULL' : ''))
+                ->setParameter('contact_methods', $contactMethods);
+        }
+
+        return $qb
+            ->orderBy('cm.description')
+            ->getQuery()
+            ->getResult();
+    }
+
+    public function getContactMethodStatsByAcademicYearWorkcenterProjectsAndMethods(
+        AcademicYear $academicYear,
+        Workcenter $workcenter = null,
+        $projects = [],
+        $contactMethods = []
+    ) {
+        $qb = $this->getEntityManager()->createQueryBuilder()
+            ->select('cm, COUNT(c)')
+            ->from(ContactMethod::class, 'cm')
+            ->join(Contact::class, 'c', 'WITH', 'c.method = cm')
+            ->leftJoin('c.method', 'm')
+            ->leftJoin('c.projects', 'p')
+            ->leftJoin('c.workcenter', 'w')
+            ->leftJoin('c.teacher', 't')
+            ->groupBy('cm')
+            ->where('cm.enabled = true AND cm.academicYear = :academic_year')
+            ->setParameter('academic_year', $academicYear)
+            ->andWhere('c.workcenter = :workcenter AND t.academicYear = :academic_year')
+            ->setParameter('workcenter', $workcenter);
 
         if ($projects !== []) {
             $noProject = false;